28.04.2020: Termino Creación Evaluaciones y Extras
Generar asignación de número de evaluación automática al abrir editor con bootstrap (OK)
-> Inicializar valores calculables en 0 (OK)
-> Ir guardando cambios de inmediato (OK)
* Cierre de Aplicativo Modal cuando se borra un audio (cierre de sub modal) (OK)
* Cierre de Aplicativo Modal cuando se borra un adjunto (cierre de sub modal) (OK)
-> validar si un archivo ha sido cargado dos veces (OK)
-> guardar audio (OK)
-> registrar en DB carga y eliminacion de adjuntos (OK)
> guardar observación de la interacción telefónica (OK)
-> Identificación de botón clickeado en la página

+++ core/AdjuntosDelete.php
@@ -1,20 +1,32 @@
 <?php
 	include("../config/Globales.php");
 	include("../config/basicos.php");
+	include("../controller/AdjuntosController.php");
+	$control = new AdjuntosController();
+
 	
 	$nombre = filter_input(INPUT_POST, ("file"));
 	$nombre = basename($nombre);
 	if($nombre == '') {
 		http_response_code(500);
 	}else{
-		if(file_exists(dirFileAttachments.$nombre)) {
-			if(unlink(dirFileAttachments.$nombre)) {
-				http_response_code(200);
-			}else{
-				http_response_code(401);
-			}
+		$obj = $control->listarPorArchivoServer($nombre);
+		if($obj == null) {
+			http_response_code(501);
 		}else{
-			http_response_code(301);
+			if($control->eliminar($obj[0])) {
+				if(file_exists(dirFileAttachments.$nombre)) {
+					if(unlink(dirFileAttachments.$nombre)) {
+						http_response_code(200);
+					}else{
+						http_response_code(401);
+					}
+				}else{
+					http_response_code(301);
+				}
+			}else{
+				http_response_code(502);
+			}
 		}
 	}
 ?>
